+++ lib/src/Utils.dart
@@ -1,17 +1,21 @@
 // ignore_for_file: file_names
 
 extension FindIntListMax on List<int> {
-  int max() =>
-      reduce((value, element) => value > element ? value : element);
-  int min() =>
-      reduce((value, element) => value < element ? value : element);
+  int max() => reduce((value, element) => value > element ? value : element);
+  int min() => reduce((value, element) => value < element ? value : element);
 }
 
 extension FindDoubleListMax on List<double> {
-  double max() =>
-      reduce((value, element) => value > element ? value : element);
-  double min() =>
-      reduce((value, element) => value < element ? value : element);
+  double max() => reduce((value, element) => element.isNaN
+      ? value
+      : value > element
+          ? value
+          : element);
+  double min() => reduce((value, element) => element.isNaN
+      ? value
+      : value < element
+          ? value
+          : element);
 }
 
 extension DateTimeFormatpr on DateTime {