+++ lib/src/widget/painter/LinePainter.dart
@@ -1,3 +1,5 @@
+import 'dart:developer';
+
 import 'package:flutter/material.dart';
 import 'package:goodlinker_chart/src/entry/ChartRuleLine.dart';
 import 'package:goodlinker_chart/src/entry/CartesianData.dart';
@@ -5,6 +7,7 @@ import 'package:goodlinker_chart/src/entry/CartesianEntry.dart';
 import 'package:goodlinker_chart/src/entry/LineChartEntry.dart';
 import 'package:goodlinker_chart/src/style/ChartStyle.dart';
 import 'package:goodlinker_chart/src/Utils.dart';
+import 'package:goodlinker_chart/src/usecases/contrast_color_calculate.dart';
 
 class LinePainter extends CustomPainter {
   final ChartStyle style;
@@ -17,10 +20,11 @@ class LinePainter extends CustomPainter {
   final int currentMiddleDisplayIndex;
   final int currentDisplayRange;
   final bool dataFeedbackMode;
+  final bool enableDrawLastDataCircle;
   final Offset lastTapLocation;
   final String Function(double value)? xAxisFormatter;
   final String Function(double value)? yAxisFormatter;
-  void Function(LineChartEntry entry, int selectIndex) dataSelectionCallback;
+  void Function(LineChartEntry? entry, int selectIndex) dataSelectionCallback;
   LinePainter({
     required this.style,
     required this.data,
@@ -28,6 +32,7 @@ class LinePainter extends CustomPainter {
     required this.currentMiddleDisplayIndex,
     required this.currentDisplayRange,
     required this.dataFeedbackMode,
+    required this.enableDrawLastDataCircle,
     required this.dataSelectionCallback,
     this.xAxisFormatter,
     this.yAxisFormatter,
@@ -63,22 +68,22 @@ class LinePainter extends CustomPainter {
 
   @override
   void paint(Canvas canvas, Size size) {
-    // _drawDebugInfo(canvas: canvas, size: size);
-    displayingEntities = _calDisplayingEntities(size);
-    _handleTap(canvas: canvas, size: size);
-
-    _drawXAxis(canvas: canvas, size: size);
-    _drawYAxis(canvas: canvas, size: size);
-    _drawArea(canvas: canvas, size: size);
-    _drawExtremeValue(canvas: canvas, size: size);
-    _drawData(canvas: canvas, size: size);
-    _drawDisplayingDataLine(canvas: canvas, size: size);
+    try {
+      // _drawDebugInfo(canvas: canvas, size: size);
+      displayingEntities = _calDisplayingEntities(size);
+      _handleTap(canvas: canvas, size: size);
+      _drawXAxis(canvas: canvas, size: size);
+      _drawYAxis(canvas: canvas, size: size);
+      _drawArea(canvas: canvas, size: size);
+      _drawExtremeValue(canvas: canvas, size: size);
+      _drawData(canvas: canvas, size: size);
+      _drawDisplayingDataLine(canvas: canvas, size: size);
+    } catch (err, s) {
+      rethrow;
+    }
   }
 
-  void _drawData({
-    required Canvas canvas,
-    required Size size,
-  }) {
+  void _drawData({required Canvas canvas, required Size size}) {
     Paint normalDataPaint = Paint()
       ..color = style.lineStyle.normalColor
       ..strokeWidth = style.lineStyle.dataLineWidth;
@@ -92,7 +97,7 @@ class LinePainter extends CustomPainter {
         final lastData = displayingEntities[index - 1];
         final pointWidth = entry.x - lastData.x;
 
-        if (lastData.y == null) {
+        if (lastData.y == null || lastData.y == double.nan) {
         } else {
           final srcOffset = Offset(lastData.x, lastData.y ?? 0);
           final tarOffset = Offset(entry.x, entry.y ?? 0);
@@ -237,10 +242,7 @@ class LinePainter extends CustomPainter {
     });
   }
 
-  void _drawExtremeValue({
-    required Canvas canvas,
-    required Size size,
-  }) {
+  void _drawExtremeValue({required Canvas canvas, required Size size}) {
     Paint maxValuePaint = Paint()
       ..color = style.lineStyle.maxValueColor
       ..strokeWidth = style.lineStyle.dataLineWidth;
@@ -249,18 +251,19 @@ class LinePainter extends CustomPainter {
       ..color = style.lineStyle.minValueColor
       ..strokeWidth = style.lineStyle.dataLineWidth;
 
-    final textStyle = TextStyle(
+    TextStyle textStyle = TextStyle(
       color: Colors.white,
       fontSize: style.xAxisStyle.axisLabelFontSize,
       // fontSize: 14,
     );
 
     ChartBaseline? thisMaxLine = maxBaseLine;
-    if (thisMaxLine != null) {
+    if (thisMaxLine != null && !thisMaxLine!.dy.isNaN) {
       if (thisMaxLine.dy <= size.height && thisMaxLine.dy >= 0) {
         final TextSpan axisTextSpan = TextSpan(
           text: maxLine?.toStringAsFixed(1),
-          style: textStyle,
+          style: textStyle.copyWith(color: ContrastColorCalculate()
+                  .determinemonochromeFontColorFromBackground(backgroundColor: maxValuePaint.color)),
         );
         final axisTextPainter = TextPainter(
           text: axisTextSpan,
@@ -270,6 +273,7 @@ class LinePainter extends CustomPainter {
           minWidth: 0,
           maxWidth: size.width,
         );
+        // drawing top extreme value y label on y axis
         canvas.drawLine(
           Offset(areaDrawingBottomRight.dx, thisMaxLine.dy),
           Offset(areaDrawingBottomRight.dx + 5, thisMaxLine.dy),
@@ -295,20 +299,33 @@ class LinePainter extends CustomPainter {
           Offset(areaDrawingBottomRight.dx + 8,
               thisMaxLine.dy - axisTextPainter.height / 2),
         );
-
-        canvas.drawLine(
-          Offset(areaDrawingAreaTopLeft.dx, thisMaxLine.dy),
-          Offset(areaDrawingBottomRight.dx, thisMaxLine.dy),
-          maxValuePaint,
-        );
+        // drawing top extreme value line
+        double dottedLineWidth = 1, dottedLineSpace = 1;
+        double startX = areaDrawingAreaTopLeft.dx;
+        while (startX < areaDrawingBottomRight.dx) {
+          canvas.drawLine(
+            Offset(startX, thisMaxLine.dy),
+            Offset(startX + dottedLineWidth, thisMaxLine.dy),
+            maxValuePaint,
+          );
+          startX += dottedLineWidth + dottedLineSpace;
+        }
+        // canvas.drawLine(
+        //   Offset(areaDrawingAreaTopLeft.dx, thisMaxLine.dy),
+        //   Offset(areaDrawingBottomRight.dx, thisMaxLine.dy),
+        //   maxValuePaint,
+        // );
       }
     }
     ChartBaseline? thisMinLine = minBaseLine;
-    if (thisMinLine != null) {
+    if (thisMinLine != null && !thisMinLine!.dy.isNaN) {
       if (thisMinLine.dy <= size.height && thisMinLine.dy >= 0) {
         final TextSpan axisTextSpan = TextSpan(
           text: minLine?.toStringAsFixed(1),
-          style: textStyle,
+          style: textStyle.copyWith(
+              color: ContrastColorCalculate()
+                  .determinemonochromeFontColorFromBackground(
+                      backgroundColor: minValuePaint.color)),
         );
         final axisTextPainter = TextPainter(
           text: axisTextSpan,
@@ -318,6 +335,8 @@ class LinePainter extends CustomPainter {
           minWidth: 0,
           maxWidth: size.width,
         );
+        // drawing bottom extreme value y label on y axis
+
         canvas.drawLine(
           Offset(areaDrawingBottomRight.dx, thisMinLine.dy),
           Offset(areaDrawingBottomRight.dx + 5, thisMinLine.dy),
@@ -343,20 +362,29 @@ class LinePainter extends CustomPainter {
           Offset(areaDrawingBottomRight.dx + 8,
               thisMinLine.dy - axisTextPainter.height / 2),
         );
+        // drawing bottom extreme value line
 
-        canvas.drawLine(
-          Offset(areaDrawingAreaTopLeft.dx, thisMinLine.dy),
-          Offset(areaDrawingBottomRight.dx, thisMinLine.dy),
-          minValuePaint,
-        );
+        double dottedLineWidth = 1, dottedLineSpace = 1;
+        double startX = areaDrawingAreaTopLeft.dx;
+        while (startX < areaDrawingBottomRight.dx) {
+          canvas.drawLine(
+            Offset(startX, thisMinLine.dy),
+            Offset(startX + dottedLineWidth, thisMinLine.dy),
+            maxValuePaint,
+          );
+          startX += dottedLineWidth + dottedLineSpace;
+        }
+
+        // canvas.drawLine(
+        //   Offset(areaDrawingAreaTopLeft.dx, thisMinLine.dy),
+        //   Offset(areaDrawingBottomRight.dx, thisMinLine.dy),
+        //   minValuePaint,
+        // );
       }
     }
   }
 
-  void _drawArea({
-    required Canvas canvas,
-    required Size size,
-  }) {
+  void _drawArea({required Canvas canvas, required Size size}) {
     Paint normalAreaPaint = Paint()
       ..color = style.lineStyle.normalColor.withAlpha(25);
     Paint maskAreaPaint = Paint()
@@ -369,21 +397,37 @@ class LinePainter extends CustomPainter {
       fontSize: style.xAxisStyle.axisLabelFontSize,
       // fontSize: 14,
     );
-    canvas.drawRect(
-      Rect.fromPoints(
-        Offset(areaDrawingAreaTopLeft.dx,
-            upperBaseline != null ? upperBaseline!.dy : 0),
-        Offset(areaDrawingBottomRight.dx,
-            lowerBaseline != null ? lowerBaseline!.dy : 0),
-      ),
-      normalAreaPaint,
-    );
+
+    try {
+      canvas.drawRect(
+        Rect.fromPoints(
+          Offset(
+            areaDrawingAreaTopLeft.dx,
+            upperBaseline != null && !upperBaseline!.dy.isNaN
+                ? upperBaseline!.dy
+                : 0,
+          ),
+          Offset(
+            areaDrawingBottomRight.dx,
+            lowerBaseline != null && !upperBaseline!.dy.isNaN
+                ? lowerBaseline!.dy
+                : areaDrawingBottomRight.dy,
+          ),
+        ),
+        normalAreaPaint,
+      );
+    } catch (err, s) {
+      rethrow;
+    }
 
     ChartBaseline? thisUpperBaseline = upperBaseline;
-    if (thisUpperBaseline != null) {
+    if (thisUpperBaseline != null && !thisUpperBaseline!.dy.isNaN) {
       final TextSpan axisTextSpan = TextSpan(
         text: upperLine?.toStringAsFixed(1),
-        style: textStyle,
+        style: textStyle.copyWith(
+              color: ContrastColorCalculate()
+                  .determinemonochromeFontColorFromBackground(
+                      backgroundColor: limitLinePaint.color)),
       );
       final axisTextPainter = TextPainter(
         text: axisTextSpan,
@@ -418,12 +462,22 @@ class LinePainter extends CustomPainter {
         Offset(areaDrawingBottomRight.dx + 8,
             thisUpperBaseline.dy - axisTextPainter.height / 2),
       );
-
-      canvas.drawLine(
-        Offset(areaDrawingAreaTopLeft.dx, thisUpperBaseline.dy),
-        Offset(areaDrawingBottomRight.dx, thisUpperBaseline.dy),
-        limitLinePaint,
-      );
+      // drawing upper base line
+      double dottedLineWidth = 1, dottedLineSpace = 1;
+      double startX = areaDrawingAreaTopLeft.dx;
+      while (startX < areaDrawingBottomRight.dx) {
+        canvas.drawLine(
+          Offset(startX, thisUpperBaseline.dy),
+          Offset(startX + dottedLineWidth, thisUpperBaseline.dy),
+          limitLinePaint,
+        );
+        startX += dottedLineWidth + dottedLineSpace;
+      }
+      // canvas.drawLine(
+      //   Offset(areaDrawingAreaTopLeft.dx, thisUpperBaseline.dy),
+      //   Offset(areaDrawingBottomRight.dx, thisUpperBaseline.dy),
+      //   limitLinePaint,
+      // );
 
       canvas.drawRRect(
         RRect.fromRectAndCorners(
@@ -438,10 +492,13 @@ class LinePainter extends CustomPainter {
       );
     }
     ChartBaseline? thisLowerBaseline = lowerBaseline;
-    if (thisLowerBaseline != null) {
+    if (thisLowerBaseline != null && !thisLowerBaseline!.dy.isNaN) {
       final TextSpan axisTextSpan = TextSpan(
         text: lowerline?.toStringAsFixed(1),
-        style: textStyle,
+        style: textStyle.copyWith(
+              color: ContrastColorCalculate()
+                  .determinemonochromeFontColorFromBackground(
+                      backgroundColor: limitLinePaint.color)),
       );
       final axisTextPainter = TextPainter(
         text: axisTextSpan,
@@ -478,12 +535,23 @@ class LinePainter extends CustomPainter {
           thisLowerBaseline.dy - axisTextPainter.height / 2,
         ),
       );
+      // drawing lower base line
+      double dottedLineWidth = 1, dottedLineSpace = 1;
+      double startX = areaDrawingAreaTopLeft.dx;
+      while (startX < areaDrawingBottomRight.dx) {
+        canvas.drawLine(
+          Offset(startX, thisLowerBaseline.dy),
+          Offset(startX + dottedLineWidth, thisLowerBaseline.dy),
+          limitLinePaint,
+        );
+        startX += dottedLineWidth + dottedLineSpace;
+      }
 
-      canvas.drawLine(
-        Offset(areaDrawingAreaTopLeft.dx, thisLowerBaseline.dy),
-        Offset(areaDrawingBottomRight.dx, thisLowerBaseline.dy),
-        limitLinePaint,
-      );
+      // canvas.drawLine(
+      //   Offset(areaDrawingAreaTopLeft.dx, thisLowerBaseline.dy),
+      //   Offset(areaDrawingBottomRight.dx, thisLowerBaseline.dy),
+      //   limitLinePaint,
+      // );
       canvas.drawRect(
         Rect.fromPoints(
           Offset(areaDrawingAreaTopLeft.dx, thisLowerBaseline.dy),
@@ -494,10 +562,7 @@ class LinePainter extends CustomPainter {
     }
   }
 
-  void _drawYAxis({
-    required Canvas canvas,
-    required Size size,
-  }) {
+  void _drawYAxis({required Canvas canvas, required Size size}) {
     final textStyle = TextStyle(
       color: Colors.black,
       fontSize: style.xAxisStyle.axisLabelFontSize,
@@ -554,10 +619,7 @@ class LinePainter extends CustomPainter {
     }
   }
 
-  void _drawXAxis({
-    required Canvas canvas,
-    required Size size,
-  }) {
+  void _drawXAxis({required Canvas canvas, required Size size}) {
     canvas.drawRRect(
       RRect.fromRectAndCorners(
         Rect.fromPoints(
@@ -589,14 +651,13 @@ class LinePainter extends CustomPainter {
         Paint()..color = style.xAxisStyle.axisBorderLineColor);
   }
 
-  _LastLabel _drawXAxisLabel({
-    required Canvas canvas,
-    required Size size,
-    required int index,
-    required Offset src,
-    required Offset tar,
-    required double xData,
-  }) {
+  _LastLabel _drawXAxisLabel(
+      {required Canvas canvas,
+      required Size size,
+      required int index,
+      required Offset src,
+      required Offset tar,
+      required double xData}) {
     final xAxisHeight = style.xAxisStyle.axisLabelFontSize + 2;
 
     final textStyle = TextStyle(
@@ -658,6 +719,8 @@ class LinePainter extends CustomPainter {
               .round()
               .clamp(0, displayingEntities.length - 1);
       dataSelectionCallback(displayingEntities[selectIndex], selectIndex);
+    } else {
+      dataSelectionCallback(null, 0);
     }
   }
 
@@ -738,7 +801,7 @@ class LinePainter extends CustomPainter {
     final List<LineChartEntry> offsetLeft = entitiesOnLeft
         .asMap()
         .map((index, entry) {
-          final disFromMid = (entitiesOnLeft.length - index - 1) * xUnit;
+          final disFromMid = (entitiesOnLeft.length - index ) * xUnit;
           final dx = dataDrawingAreaTopLeft.dx + middleXPos - disFromMid;
           final dy = entry?.y == null
               ? null