+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.0.17
+    bugfix: chart entities overlapping.
+## 0.0.16
+    add adaptive font color and dotted line style.
+
+## 0.0.15
+    remove minimal height for BarChart.
+
+## 0.0.14
+    add data support NaN.
+
+## 0.0.13
+    add minimal height for BarChart.
+
+## 0.0.12
+    add parameter to controll wether or not to draw circle on last data in SimpleLineChart and LineChart (which has no effect due to they don't draw it in the first place.).
+
+## 0.0.11
+    Remove usage of GestureDetector in SimpleBarChart and SimpleLineChart. They are now not interactive with, the enableXScale and the enableXScroll property will be removed in future.
+
+## 0.0.10
+    adjust data feedback to nullable
+## 0.0.9 
+    formatting code.
 ## 0.0.8
     avoid unnecessary_statements